Adds answer to 1.1 in Cracking the Code
The question asks the user to implement an algorithm that determines
if a string is unique without using any data structures.

I decided to play around with Swift. If you would like to play with the code yourself
you can open this file up directly with the latest version of Xcode in “Playground” mode.